+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/presentation/ChatCompositeActivity.kt
@@ -16,7 +16,10 @@ class ChatCompositeActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         addContentView(
             ChatView(this, instanceId),
-            ViewGroup.LayoutParams(FrameLayout.LayoutParams.MATCH_PARENT, FrameLayout.LayoutParams.MATCH_PARENT)
+            ViewGroup.LayoutParams(
+                FrameLayout.LayoutParams.MATCH_PARENT,
+                FrameLayout.LayoutParams.MATCH_PARENT
+            )
         )
     }
