+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/presentation/ui/reduxviewmodel/ReduxViewModel.kt
@@ -1,4 +1,4 @@
-package com.azure.android.communication.ui.chat.presentation.ui.redux_view_model
+package com.azure.android.communication.ui.chat.presentation.ui.reduxviewmodel
 
 import com.azure.android.communication.ui.chat.redux.AppStore
 import kotlinx.coroutines.CoroutineScope
