+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/presentation/ui/viewmodel/MessageViewModel.kt
@@ -1,16 +1,22 @@
-package com.azure.android.communication.ui.chat.presentation.ui.view_model
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.android.communication.ui.chat.presentation.ui.viewmodel
 
 import com.azure.android.communication.ui.chat.models.MessageInfoModel
 
 internal class MessageViewModel(val message: MessageInfoModel)
 
-internal fun List<MessageInfoModel>.toViewModelList() = InfoModelToViewModelAdapter(this) as List<MessageViewModel>
+internal fun List<MessageInfoModel>.toViewModelList() =
+    InfoModelToViewModelAdapter(this) as List<MessageViewModel>
 
-private class InfoModelToViewModelAdapter(private val messages: List<MessageInfoModel>) : List<MessageViewModel> {
+private class InfoModelToViewModelAdapter(private val messages: List<MessageInfoModel>) :
+    List<MessageViewModel> {
     override val size = messages.size
     override fun contains(element: MessageViewModel) = messages.contains(element.message)
 
-    override fun containsAll(elements: Collection<MessageViewModel>) = messages.containsAll(elements.map { it.message })
+    override fun containsAll(elements: Collection<MessageViewModel>) =
+        messages.containsAll(elements.map { it.message })
 
     override fun get(index: Int) = MessageViewModel(messages[index])
